# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "basePath": "/v1",
    "paths": {
        "/api/chat": {
            "post": {
                "description": "チャットを行うAPI",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "チャットAPI",
                "operationId": "post-chat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "セッションID",
                        "name": "session_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "チャットメッセージ",
                        "name": "chat_message",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/chat-start": {
            "post": {
                "description": "最初の画像を送信してチャットを開始するAPI",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "最初の画像送信用チャットAPI",
                "operationId": "post-chat-start",
                "parameters": [
                    {
                        "type": "string",
                        "description": "セッションID",
                        "name": "session_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "画像データ",
                        "name": "base64_image",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/products": {
            "get": {
                "description": "商品一覧を取得するAPI",
                "produces": [
                    "application/json"
                ],
                "summary": "商品一覧取得API",
                "operationId": "get-products",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/main.ProductsJson"
                            }
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.ProductsJson": {
            "type": "object",
            "properties": {
                "image": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "price": {
                    "type": "number"
                }
            }
        }
    }
}